+++ packages/taro-cli/src/weapp.js
@@ -4,6 +4,7 @@ const path = require('path')
 const chalk = require('chalk')
 const chokidar = require('chokidar')
 const wxTransformer = require('@tarojs/transformer-wx')
+const babel = require('babel-core')
 const traverse = require('babel-traverse').default
 const t = require('babel-types')
 const generate = require('babel-generator').default
@@ -23,7 +24,6 @@ const babylonConfig = require('./config/babylon')
 const browserList = require('./config/browser_list')
 const defaultUglifyConfig = require('./config/uglify')
 const defaultBabelConfig = require('./config/babel')
-const defaultTSConfig = require('./config/tsconfig.json')
 const astConvert = require('./util/ast_convert')
 
 const appPath = process.cwd()
@@ -47,7 +47,6 @@ const appOutput = typeof weappConf.appOutput === 'boolean' ? weappConf.appOutput
 
 const notExistNpmList = []
 const taroJsFramework = '@tarojs/taro'
-const taroWeappFramework = '@tarojs/taro-weapp'
 const taroJsComponents = '@tarojs/components'
 const taroJsRedux = '@tarojs/redux'
 let appConfig = {}
@@ -62,6 +61,8 @@ let isBuildingScripts = {}
 let isBuildingStyles = {}
 let isCopyingFiles = {}
 let isProduction = false
+let buildAdapter = Util.BUILD_TYPES.WEAPP
+let outputFilesTypes = Util.MINI_APP_FILES[buildAdapter]
 
 const NODE_MODULES = 'node_modules'
 const NODE_MODULES_REG = /(.*)node_modules/
@@ -104,17 +105,35 @@ function getExactedNpmFilePath (npmName, filePath) {
   }
 }
 
-function processIfTaroEnv (astPath, node, a, b) {
-  if (node[a].value !== Util.BUILD_TYPES.WEAPP) {
-    const consequentSibling = astPath.getSibling('consequent')
-    consequentSibling.set('body', [])
-  } else {
-    const alternateSibling = astPath.getSibling('alternate')
-    if (alternateSibling.node) {
-      alternateSibling.set('body', [])
-    }
+function traverseObjectNode (node, obj) {
+  if (node.type === 'ClassProperty' || node.type === 'ObjectProperty') {
+    const properties = node.value.properties
+    obj = {}
+    properties.forEach(p => {
+      const key = t.isIdentifier(p.key) ? p.key.name : p.key.value
+      obj[key] = traverseObjectNode(p.value)
+    })
+    return obj
   }
-  node[b] = t.stringLiteral(Util.BUILD_TYPES.WEAPP)
+  if (node.type === 'ObjectExpression') {
+    const properties = node.properties
+    obj = {}
+    properties.forEach(p => {
+      let key = t.isIdentifier(p.key) ? p.key.name : p.key.value
+      if (Util.CONFIG_MAP[buildAdapter][key]) {
+        key = Util.CONFIG_MAP[buildAdapter][key]
+      }
+      obj[key] = traverseObjectNode(p.value)
+    })
+    return obj
+  }
+  if (node.type === 'ArrayExpression') {
+    return node.elements.map(item => traverseObjectNode(item))
+  }
+  if (node.type === 'NullLiteral') {
+    return null
+  }
+  return node.value
 }
 
 function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip = false) {
@@ -125,36 +144,19 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
   let configObj = {}
   let componentClassName = null
   let taroJsReduxConnect = null
-  function traverseObjectNode (node, obj) {
-    if (node.type === 'ClassProperty' || node.type === 'ObjectProperty') {
-      const properties = node.value.properties
-      obj = {}
-      properties.forEach(p => {
-        const key = t.isIdentifier(p.key) ? p.key.name : p.key.value
-        obj[key] = traverseObjectNode(p.value)
-      })
-      return obj
-    }
-    if (node.type === 'ObjectExpression') {
-      const properties = node.properties
-      obj = {}
-      properties.forEach(p => {
-        const key = t.isIdentifier(p.key) ? p.key.name : p.key.value
-        obj[key] = traverseObjectNode(p.value)
-      })
-      return obj
-    }
-    if (node.type === 'ArrayExpression') {
-      return node.elements.map(item => traverseObjectNode(item))
-    }
-    if (node.type === 'NullLiteral') {
-      return null
-    }
-    return node.value
-  }
+  let taroMiniAppFramework = `@tarojs/taro-${buildAdapter}`
   let taroImportDefaultName
   let needExportDefault = false
   let exportTaroReduxConnected = null
+  const constantsReplaceList = Object.assign({
+    'process.env.TARO_ENV': Util.BUILD_TYPES.WEAPP
+  }, Util.generateEnvList(projectConfig.env || {}), Util.generateConstantsList(projectConfig.defineConstants || {}))
+  ast = babel.transformFromAst(ast, '', {
+    plugins: [
+      [require('babel-plugin-danger-remove-unused-import'), { ignore: ['@tarojs/taro', 'react', 'nervjs'] }],
+      [require('babel-plugin-transform-define').default, constantsReplaceList]
+    ]
+  }).ast
   traverse(ast, {
     ClassDeclaration (astPath) {
       const node = astPath.node
@@ -246,21 +248,6 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
       }
     },
 
-    IfStatement (astPath) {
-      astPath.traverse({
-        BinaryExpression (astPath) {
-          const node = astPath.node
-          const left = node.left
-          const right = node.right
-          if (generate(left).code === 'process.env.TARO_ENV') {
-            processIfTaroEnv(astPath, node, 'right', 'left')
-          } else if (generate(right).code === 'process.env.TARO_ENV') {
-            processIfTaroEnv(astPath, node, 'left', 'right')
-          }
-        }
-      })
-    },
-
     ImportDeclaration (astPath) {
       const node = astPath.node
       const source = node.source
@@ -291,7 +278,7 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
               if (defaultSpecifier) {
                 taroImportDefaultName = defaultSpecifier
               }
-              value = taroWeappFramework
+              value = taroMiniAppFramework
             } else if (value === taroJsRedux) {
               specifiers.forEach(item => {
                 if (item.type === 'ImportSpecifier') {
@@ -340,7 +327,7 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
             } else {
               if (value === taroJsFramework && id.type === 'Identifier') {
                 taroImportDefaultName = id.name
-                value = taroWeappFramework
+                value = taroMiniAppFramework
               } else if (value === taroJsRedux) {
                 const declarations = node.declarations
                 declarations.forEach(item => {
@@ -693,7 +680,7 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
           const exportDefault = template(`export default ${exportVariableName}`, babylonConfig)()
           node.body.push(exportDefault)
         }
-        const taroWeappFrameworkPath = !npmSkip ? getExactedNpmFilePath(taroWeappFramework, filePath) : taroWeappFramework
+        const taroMiniAppFrameworkPath = !npmSkip ? getExactedNpmFilePath(taroMiniAppFramework, filePath) : taroMiniAppFramework
         switch (type) {
           case PARSE_AST_TYPE.ENTRY:
             const pxTransformConfig = {
@@ -702,14 +689,18 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
             if (projectConfig.hasOwnProperty(DEVICE_RATIO)) {
               pxTransformConfig[DEVICE_RATIO] = projectConfig.deviceRatio
             }
-            node.body.push(template(`App(require('${taroWeappFrameworkPath}').default.createApp(${exportVariableName}))`, babylonConfig)())
+            node.body.push(template(`App(require('${taroMiniAppFrameworkPath}').default.createApp(${exportVariableName}))`, babylonConfig)())
             node.body.push(template(`Taro.initPxTransform(${JSON.stringify(pxTransformConfig)})`, babylonConfig)())
             break
           case PARSE_AST_TYPE.PAGE:
-            node.body.push(template(`Component(require('${taroWeappFrameworkPath}').default.createComponent(${exportVariableName}, true))`, babylonConfig)())
+            if (buildAdapter === Util.BUILD_TYPES.WEAPP) {
+              node.body.push(template(`Component(require('${taroMiniAppFrameworkPath}').default.createComponent(${exportVariableName}, true))`, babylonConfig)())
+            } else {
+              node.body.push(template(`Page(require('${taroMiniAppFrameworkPath}').default.createComponent(${exportVariableName}, true))`, babylonConfig)())
+            }
             break
           case PARSE_AST_TYPE.COMPONENT:
-            node.body.push(template(`Component(require('${taroWeappFrameworkPath}').default.createComponent(${exportVariableName}))`, babylonConfig)())
+            node.body.push(template(`Component(require('${taroMiniAppFrameworkPath}').default.createComponent(${exportVariableName}))`, babylonConfig)())
             break
           default:
             break
@@ -718,7 +709,7 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
     }
   })
   return {
-    code: unescape(generate(ast).code.replace(/\\u/g, '%u')),
+    code: generate(ast).code,
     styleFiles,
     scriptFiles,
     jsonFiles,
@@ -731,6 +722,14 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath, npmSkip =
 function parseComponentExportAst (ast, componentName, componentPath, componentType) {
   let componentRealPath = null
   let importExportName
+  const constantsReplaceList = Object.assign({
+    'process.env.TARO_ENV': Util.BUILD_TYPES.WEAPP
+  }, Util.generateEnvList(projectConfig.env || {}), Util.generateConstantsList(projectConfig.defineConstants || {}))
+  ast = babel.transformFromAst(ast, '', {
+    plugins: [
+      [require('babel-plugin-transform-define').default, constantsReplaceList]
+    ]
+  }).ast
   traverse(ast, {
     ExportNamedDeclaration (astPath) {
       const node = astPath.node
@@ -759,29 +758,6 @@ function parseComponentExportAst (ast, componentName, componentPath, componentTy
       if (componentType === 'default') {
         importExportName = declaration.name
       }
-    },
-
-    IfStatement (astPath) {
-      astPath.traverse({
-        BinaryExpression (astPath) {
-          const node = astPath.node
-          const left = node.left
-          if (generate(left).code === 'process.env.TARO_ENV' &&
-            node.right.value === Util.BUILD_TYPES.WEAPP) {
-            const consequentSibling = astPath.getSibling('consequent')
-            consequentSibling.traverse({
-              CallExpression (astPath) {
-                if (astPath.get('callee').isIdentifier({ name : 'require'})) {
-                  const arg = astPath.get('arguments')[0]
-                  if (t.isStringLiteral(arg.node)) {
-                    componentRealPath = Util.resolveScriptPath(path.resolve(path.dirname(componentPath), arg.node.value))
-                  }
-                }
-              }
-            })
-          }
-        }
-      })
     },
 
     CallExpression (astPath) {
@@ -902,30 +878,46 @@ function copyFilesFromSrcToOutput (files) {
   })
 }
 
-async function compileScriptFile (filePath, content) {
-  const babelConfig = Object.assign({}, pluginsConfig.babel, defaultBabelConfig)
-  const tsConfig = Object.assign({}, pluginsConfig.typescript, defaultTSConfig)
-  if (Util.REG_TYPESCRIPT.test(filePath)) {
-    const compileTSRes = await npmProcess.callPlugin('typescript', content, entryFilePath, tsConfig)
-    if (compileTSRes && compileTSRes.outputText) {
-      content = compileTSRes.outputText
-    }
+const babelConfig = _.mergeWith(defaultBabelConfig, pluginsConfig.babel, (objValue, srcValue) => {
+  if (Array.isArray(objValue)) {
+    return Array.from(new Set(objValue.concat(srcValue)))
   }
+})
+
+async function compileScriptFile (content) {
   const compileScriptRes = await npmProcess.callPlugin('babel', content, entryFilePath, babelConfig)
   return compileScriptRes.code
 }
 
 function buildProjectConfig () {
-  const projectConfigPath = path.join(appPath, 'project.config.json')
-  if (!fs.existsSync(projectConfigPath)) {
-    return
+  let projectConfigFileName = ''
+  if (buildAdapter === Util.BUILD_TYPES.WEAPP) {
+    projectConfigFileName = 'project.config.json'
+    const projectConfigPath = path.join(appPath, projectConfigFileName)
+    if (!fs.existsSync(projectConfigPath)) {
+      return
+    }
+    const origProjectConfig = fs.readJSONSync(projectConfigPath)
+    fs.ensureDirSync(outputDir)
+    fs.writeFileSync(
+      path.join(outputDir, projectConfigFileName),
+      JSON.stringify(Object.assign({}, origProjectConfig, { miniprogramRoot: './' }), null, 2)
+    )
+  } else if (buildAdapter === Util.BUILD_TYPES.SWAN) {
+    projectConfigFileName = 'project.swan.json'
+    const projectConfigObj = {
+      appid: 'testappid',
+      setting: {
+        urlCheck: false
+      }
+    }
+    fs.ensureDirSync(outputDir)
+    fs.writeFileSync(
+      path.join(outputDir, projectConfigFileName),
+      JSON.stringify(projectConfigObj, null, 2)
+    )
   }
-  const origProjectConfig = fs.readJSONSync(projectConfigPath)
-  fs.writeFileSync(
-    path.join(outputDir, 'project.config.json'),
-    JSON.stringify(Object.assign({}, origProjectConfig, { miniprogramRoot: './' }), null, 2)
-  )
-  Util.printLog(Util.pocessTypeEnum.GENERATE, '工具配置', `${outputDirName}/project.config.json`)
+  Util.printLog(Util.pocessTypeEnum.GENERATE, '工具配置', `${outputDirName}/${projectConfigFileName}`)
 }
 
 async function buildEntry () {
@@ -937,14 +929,13 @@ async function buildEntry () {
       sourcePath: entryFilePath,
       outputPath: outputEntryFilePath,
       isApp: true,
-      isTyped: Util.REG_TYPESCRIPT.test(entryFilePath)
+      isTyped: Util.REG_TYPESCRIPT.test(entryFilePath),
+      adapter: buildAdapter
     })
     // app.js的template忽略
     const res = parseAst(PARSE_AST_TYPE.ENTRY, transformResult.ast, [], entryFilePath, outputEntryFilePath)
     let resCode = res.code
-    resCode = await compileScriptFile(entryFilePath, resCode)
-    resCode = Util.replaceContentEnv(resCode, projectConfig.env || {})
-    resCode = Util.replaceContentConstants(resCode, projectConfig.defineConstants || {})
+    resCode = await compileScriptFile(resCode)
     if (isProduction) {
       const uglifyPluginConfig = pluginsConfig.uglify || { enable: true }
       if (uglifyPluginConfig.enable) {
@@ -970,8 +961,8 @@ async function buildEntry () {
     }
     // 编译样式文件
     if (Util.isDifferentArray(fileDep['style'], res.styleFiles) && appOutput) {
-      await compileDepStyles(path.join(outputDir, 'app.wxss'), res.styleFiles, false)
-      Util.printLog(Util.pocessTypeEnum.GENERATE, '入口样式', `${outputDirName}/app.wxss`)
+      await compileDepStyles(path.join(outputDir, `app${outputFilesTypes.STYLE}`), res.styleFiles, false)
+      Util.printLog(Util.pocessTypeEnum.GENERATE, '入口样式', `${outputDirName}/app${outputFilesTypes.STYLE}`)
     }
     // 拷贝依赖文件
     if (Util.isDifferentArray(fileDep['json'], res.jsonFiles)) {
@@ -1035,6 +1026,35 @@ async function buildPages () {
   await Promise.all(pagesPromises)
 }
 
+function processNativeWxml (componentWXMLPath, componentWXMLContent, outputComponentWXMLPath) {
+  let wxmlContent
+  let needCopy = true
+  if (componentWXMLPath && fs.existsSync(componentWXMLPath)) {
+    wxmlContent = fs.readFileSync(componentWXMLPath).toString()
+  } else {
+    needCopy = false
+    wxmlContent = componentWXMLContent
+  }
+  const importWxmlPathList = []
+  let regResult
+  while ((regResult = Util.REG_WXML_IMPORT.exec(wxmlContent)) != null)  {
+    importWxmlPathList.push(regResult[2] || regResult[3])
+  }
+  if (importWxmlPathList.length) {
+    importWxmlPathList.forEach(item => {
+      const itemPath = path.resolve(componentWXMLPath, '..', item)
+      if (fs.existsSync(itemPath)) {
+        const outputItemPath = itemPath.replace(sourceDir, outputDir)
+        processNativeWxml(itemPath, null, outputItemPath)
+      }
+    })
+  }
+  if (componentWXMLPath === outputComponentWXMLPath || !needCopy) {
+    return
+  }
+  copyFileSync(componentWXMLPath, outputComponentWXMLPath)
+}
+
 function transfromNativeComponents (configFile, componentConfig) {
   const usingComponents = componentConfig.usingComponents
   if (usingComponents && !Util.isEmptyObject(usingComponents)) {
@@ -1046,10 +1066,10 @@ function transfromNativeComponents (configFile, componentConfig) {
         return
       }
       const componentJSPath = Util.resolveScriptPath(path.resolve(path.dirname(configFile), componentPath))
-      const componentJSONPath = componentJSPath.replace(path.extname(componentJSPath), '.json')
-      const componentWXMLPath = componentJSPath.replace(path.extname(componentJSPath), '.wxml')
-      const componentWXSSPath = componentJSPath.replace(path.extname(componentJSPath), '.wxss')
-      const outputComponentJSPath = componentJSPath.replace(sourceDir, outputDir).replace(path.extname(componentJSPath), '.js')
+      const componentJSONPath = componentJSPath.replace(path.extname(componentJSPath), outputFilesTypes.CONFIG)
+      const componentWXMLPath = componentJSPath.replace(path.extname(componentJSPath), outputFilesTypes.TEMPL)
+      const componentWXSSPath = componentJSPath.replace(path.extname(componentJSPath), outputFilesTypes.STYLE)
+      const outputComponentJSPath = componentJSPath.replace(sourceDir, outputDir).replace(path.extname(componentJSPath), outputFilesTypes.SCRIPT)
       if (fs.existsSync(componentJSPath)) {
         const componentJSContent = fs.readFileSync(componentJSPath).toString()
         if (componentJSContent.indexOf(taroJsFramework) >= 0 && !fs.existsSync(componentWXMLPath)) {
@@ -1060,16 +1080,16 @@ function transfromNativeComponents (configFile, componentConfig) {
         return Util.printLog(Util.pocessTypeEnum.ERROR, '编译错误', `原生组件文件 ${componentJSPath} 不存在！`)
       }
       if (fs.existsSync(componentWXMLPath)) {
-        const outputComponentWXMLPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), '.wxml')
-        copyFileSync(componentWXMLPath, outputComponentWXMLPath)
+        const outputComponentWXMLPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), outputFilesTypes.TEMPL)
+        processNativeWxml(componentWXMLPath, null, outputComponentWXMLPath)
       }
       if (fs.existsSync(componentWXSSPath)) {
-        const outputComponentWXSSPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), '.wxss')
+        const outputComponentWXSSPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), outputFilesTypes.STYLE)
         await compileDepStyles(outputComponentWXSSPath, [componentWXSSPath], true)
       }
       if (fs.existsSync(componentJSONPath)) {
         const componentJSON = require(componentJSONPath)
-        const outputComponentJSONPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), '.json')
+        const outputComponentJSONPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), outputFilesTypes.CONFIG)
         copyFileSync(componentJSONPath, outputComponentJSONPath)
         transfromNativeComponents(componentJSONPath, componentJSON)
       }
@@ -1087,16 +1107,16 @@ async function buildSinglePage (page) {
     return
   }
   const pageJsContent = fs.readFileSync(pageJs).toString()
-  const outputPageJSPath = pageJs.replace(sourceDir, outputDir).replace(path.extname(pageJs), '.js')
+  const outputPageJSPath = pageJs.replace(sourceDir, outputDir).replace(path.extname(pageJs), outputFilesTypes.SCRIPT)
   const outputPagePath = path.dirname(outputPageJSPath)
-  const outputPageJSONPath = outputPageJSPath.replace(path.extname(outputPageJSPath), '.json')
-  const outputPageWXMLPath = outputPageJSPath.replace(path.extname(outputPageJSPath), '.wxml')
-  const outputPageWXSSPath = outputPageJSPath.replace(path.extname(outputPageJSPath), '.wxss')
+  const outputPageJSONPath = outputPageJSPath.replace(path.extname(outputPageJSPath), outputFilesTypes.CONFIG)
+  const outputPageWXMLPath = outputPageJSPath.replace(path.extname(outputPageJSPath), outputFilesTypes.TEMPL)
+  const outputPageWXSSPath = outputPageJSPath.replace(path.extname(outputPageJSPath), outputFilesTypes.STYLE)
   // 判断是不是小程序原生代码页面
-  const pageWXMLPath = pageJs.replace(path.extname(pageJs), '.wxml')
+  const pageWXMLPath = pageJs.replace(path.extname(pageJs), outputFilesTypes.TEMPL)
   if (fs.existsSync(pageWXMLPath) && pageJsContent.indexOf(taroJsFramework) < 0) {
-    const pageJSONPath = pageJs.replace(path.extname(pageJs), '.json')
-    const pageWXSSPath = pageJs.replace(path.extname(pageJs), '.wxss')
+    const pageJSONPath = pageJs.replace(path.extname(pageJs), outputFilesTypes.CONFIG)
+    const pageWXSSPath = pageJs.replace(path.extname(pageJs), outputFilesTypes.STYLE)
     if (fs.existsSync(pageJSONPath)) {
       const pageJSON = require(pageJSONPath)
       copyFileSync(pageJSONPath, outputPageJSONPath)
@@ -1115,14 +1135,13 @@ async function buildSinglePage (page) {
       sourcePath: pageJs,
       outputPath: outputPageJSPath,
       isRoot: true,
-      isTyped: Util.REG_TYPESCRIPT.test(pageJs)
+      isTyped: Util.REG_TYPESCRIPT.test(pageJs),
+      adapter: buildAdapter
     })
     const pageDepComponents = transformResult.components
     const res = parseAst(PARSE_AST_TYPE.PAGE, transformResult.ast, pageDepComponents, pageJs, outputPageJSPath)
     let resCode = res.code
-    resCode = await compileScriptFile(pageJs, resCode)
-    resCode = Util.replaceContentEnv(resCode, projectConfig.env || {})
-    resCode = Util.replaceContentConstants(resCode, projectConfig.defineConstants || {})
+    resCode = await compileScriptFile(resCode)
     if (isProduction) {
       const uglifyPluginConfig = pluginsConfig.uglify || { enable: true }
       if (uglifyPluginConfig.enable) {
@@ -1187,18 +1206,19 @@ async function buildSinglePage (page) {
       })
     }
     fs.writeFileSync(outputPageJSONPath, JSON.stringify(_.merge({}, buildUsingComponents(pageDepComponents), res.configObj), null, 2))
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '页面JSON', `${outputDirName}/${page}.json`)
+    Util.printLog(Util.pocessTypeEnum.GENERATE, '页面配置', `${outputDirName}/${page}${outputFilesTypes.CONFIG}`)
     fs.writeFileSync(outputPageJSPath, resCode)
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '页面JS', `${outputDirName}/${page}.js`)
+    Util.printLog(Util.pocessTypeEnum.GENERATE, '页面逻辑', `${outputDirName}/${page}${outputFilesTypes.SCRIPT}`)
     fs.writeFileSync(outputPageWXMLPath, transformResult.template)
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '页面WXML', `${outputDirName}/${page}.wxml`)
+    processNativeWxml(outputPageWXMLPath.replace(outputDir, sourceDir), transformResult.template, outputPageWXMLPath)
+    Util.printLog(Util.pocessTypeEnum.GENERATE, '页面模板', `${outputDirName}/${page}${outputFilesTypes.TEMPL}`)
     // 编译依赖的脚本文件
     if (Util.isDifferentArray(fileDep['script'], res.scriptFiles)) {
       compileDepScripts(res.scriptFiles)
     }
     // 编译样式文件
     if (Util.isDifferentArray(fileDep['style'], res.styleFiles) || Util.isDifferentArray(depComponents[pageJs], pageDepComponents)) {
-      Util.printLog(Util.pocessTypeEnum.GENERATE, '页面WXSS', `${outputDirName}/${page}.wxss`)
+      Util.printLog(Util.pocessTypeEnum.GENERATE, '页面样式', `${outputDirName}/${page}${outputFilesTypes.STYLE}`)
       const depStyleList = getDepStyleList(outputPageWXSSPath, buildDepComponentsResult)
       wxssDepTree[outputPageWXSSPath] = depStyleList
       await compileDepStyles(outputPageWXSSPath, res.styleFiles, false)
@@ -1225,8 +1245,22 @@ async function buildSinglePage (page) {
 async function processStyleWithPostCSS (styleObj) {
   const useModuleConf = weappConf.module || {}
   const customPostcssConf = useModuleConf.postcss || {}
-  const customPxtransformConf = customPostcssConf.pxtransform || {}
-  const customUrlConf = customPostcssConf.url || {}
+  const customPxtransformConf =  Object.assign({
+    enable: true,
+    config: {}
+  }, customPostcssConf.pxtransform || {})
+  const customUrlConf =  Object.assign({
+    enable: true,
+    config: {
+      limit: 10240
+    }
+  }, customPostcssConf.url || {})
+  const customAutoprefixerConf = Object.assign({
+    enable: true,
+    config: {
+      browsers: browserList
+    }
+  }, customPostcssConf.autoprefixer || {})
   const postcssPxtransformOption = {
     designWidth: projectConfig.designWidth || 750,
     platform: 'weapp'
@@ -1235,15 +1269,16 @@ async function processStyleWithPostCSS (styleObj) {
   if (projectConfig.hasOwnProperty(DEVICE_RATIO)) {
     postcssPxtransformOption[DEVICE_RATIO] = projectConfig.deviceRatio
   }
-  const cssUrlConf = Object.assign({ limit: 10240, enable: true }, customUrlConf)
-  const maxSize = Math.round(cssUrlConf.limit / 1024)
-  const processors = [
-    autoprefixer({ browsers: browserList }),
-    pxtransform(Object.assign(
-      postcssPxtransformOption,
-      customPxtransformConf
-    ))
-  ]
+  const cssUrlConf = Object.assign({ limit: 10240 }, customUrlConf)
+  const maxSize = Math.round((customUrlConf.config.limit || cssUrlConf.limit) / 1024)
+  const postcssPxtransformConf = Object.assign({}, postcssPxtransformOption, customPxtransformConf, customPxtransformConf.config)
+  const processors = []
+  if (customAutoprefixerConf.enable) {
+    processors.push(autoprefixer(customAutoprefixerConf.config))
+  }
+  if (customPxtransformConf.enable) {
+    processors.push(pxtransform(postcssPxtransformConf))
+  }
   if (cssUrlConf.enable) {
     processors.push(cssUrlParse({
       url: 'inline',
@@ -1257,6 +1292,17 @@ async function processStyleWithPostCSS (styleObj) {
   return postcssResult.css
 }
 
+function compileImportStyles (filePath, importStyles) {
+  if (importStyles.length) {
+    importStyles.forEach(async importItem => {
+      const importFilePath = path.resolve(filePath, '..', importItem)
+      if (fs.existsSync(importFilePath)) {
+        await compileDepStyles(importFilePath.replace(sourceDir, outputDir), [importFilePath])
+      }
+    })
+  }
+}
+
 function compileDepStyles (outputFilePath, styleFiles, isComponent) {
   if (isBuildingStyles[outputFilePath]) {
     return Promise.resolve({})
@@ -1267,17 +1313,18 @@ function compileDepStyles (outputFilePath, styleFiles, isComponent) {
     const fileExt = path.extname(filePath)
     const pluginName = Util.FILE_PROCESSOR_MAP[fileExt]
     const fileContent = fs.readFileSync(filePath).toString()
-    const cssImportsRes = Util.processWxssImports(fileContent)
+    const cssImportsRes = Util.processStyleImports(fileContent, buildAdapter)
+    compileImportStyles(filePath, cssImportsRes.imports)
     if (pluginName) {
       return npmProcess.callPlugin(pluginName, cssImportsRes.content, filePath, pluginsConfig[pluginName] || {})
         .then(res => ({
-          css: cssImportsRes.wxss.join('\n') + '\n' + res.css,
+          css: cssImportsRes.style.join('\n') + '\n' + res.css,
           filePath
         }))
     }
     return new Promise(resolve => {
       resolve({
-        css: cssImportsRes.wxss.join('\n') + '\n' + cssImportsRes.content,
+        css: cssImportsRes.style.join('\n') + '\n' + cssImportsRes.content,
         filePath
       })
     })
@@ -1382,10 +1429,10 @@ async function buildSingleComponent (componentObj, buildConfig = {}) {
   outputComponentShowPath = outputComponentShowPath.replace(path.extname(outputComponentShowPath), '')
   Util.printLog(Util.pocessTypeEnum.COMPILE, '组件文件', componentShowPath)
   const componentContent = fs.readFileSync(component).toString()
-  const outputComponentJSPath = component.replace(sourceDirPath, buildConfig.outputDir || buildOutputDir).replace(path.extname(component), '.js')
-  const outputComponentWXMLPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), '.wxml')
-  const outputComponentWXSSPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), '.wxss')
-  const outputComponentJSONPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), '.json')
+  const outputComponentJSPath = component.replace(sourceDirPath, buildConfig.outputDir || buildOutputDir).replace(path.extname(component), outputFilesTypes.SCRIPT)
+  const outputComponentWXMLPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), outputFilesTypes.TEMPL)
+  const outputComponentWXSSPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), outputFilesTypes.STYLE)
+  const outputComponentJSONPath = outputComponentJSPath.replace(path.extname(outputComponentJSPath), outputFilesTypes.CONFIG)
   if (hasBeenBuiltComponents.indexOf(component) < 0) {
     hasBeenBuiltComponents.push(component)
   }
@@ -1422,14 +1469,13 @@ async function buildSingleComponent (componentObj, buildConfig = {}) {
       outputPath: outputComponentJSPath,
       isRoot: false,
       isTyped: Util.REG_TYPESCRIPT.test(component),
-      isNormal: false
+      isNormal: false,
+      adapter: buildAdapter
     })
     const componentDepComponents = transformResult.components
     const res = parseAst(PARSE_AST_TYPE.COMPONENT, transformResult.ast, componentDepComponents, component, outputComponentJSPath, buildConfig.npmSkip)
     let resCode = res.code
-    resCode = await compileScriptFile(component, resCode)
-    resCode = Util.replaceContentEnv(resCode, projectConfig.env || {})
-    resCode = Util.replaceContentConstants(resCode, projectConfig.defineConstants || {})
+    resCode = await compileScriptFile(resCode)
     fs.ensureDirSync(path.dirname(outputComponentJSPath))
     if (isProduction) {
       const uglifyPluginConfig = pluginsConfig.uglify || { enable: true }
@@ -1495,18 +1541,19 @@ async function buildSingleComponent (componentObj, buildConfig = {}) {
       })
     }
     fs.writeFileSync(outputComponentJSONPath, JSON.stringify(_.merge({}, buildUsingComponents(componentDepComponents, true), res.configObj), null, 2))
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '组件JSON', `${outputDirName}/${outputComponentShowPath}.json`)
+    Util.printLog(Util.pocessTypeEnum.GENERATE, '组件配置', `${outputDirName}/${outputComponentShowPath}${outputFilesTypes.CONFIG}`)
     fs.writeFileSync(outputComponentJSPath, resCode)
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '组件JS', `${outputDirName}/${outputComponentShowPath}.js`)
+    Util.printLog(Util.pocessTypeEnum.GENERATE, '组件逻辑', `${outputDirName}/${outputComponentShowPath}${outputFilesTypes.SCRIPT}`)
     fs.writeFileSync(outputComponentWXMLPath, transformResult.template)
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '组件WXML', `${outputDirName}/${outputComponentShowPath}.wxml`)
+    processNativeWxml(outputComponentWXMLPath.replace(outputDir, sourceDir), transformResult.template, outputComponentWXMLPath)
+    Util.printLog(Util.pocessTypeEnum.GENERATE, '组件模板', `${outputDirName}/${outputComponentShowPath}${outputFilesTypes.TEMPL}`)
     // 编译依赖的脚本文件
     if (Util.isDifferentArray(fileDep['script'], res.scriptFiles)) {
       compileDepScripts(res.scriptFiles)
     }
     // 编译样式文件
     if (Util.isDifferentArray(fileDep['style'], res.styleFiles) || Util.isDifferentArray(depComponents[component], componentDepComponents)) {
-      Util.printLog(Util.pocessTypeEnum.GENERATE, '组件WXSS', `${outputDirName}/${outputComponentShowPath}.wxss`)
+      Util.printLog(Util.pocessTypeEnum.GENERATE, '组件样式', `${outputDirName}/${outputComponentShowPath}${outputFilesTypes.STYLE}`)
       const depStyleList = getDepStyleList(outputComponentWXSSPath, buildDepComponentsResult)
       wxssDepTree[outputComponentWXSSPath] = depStyleList
       await compileDepStyles(outputComponentWXSSPath, res.styleFiles, true)
@@ -1572,10 +1619,8 @@ function compileDepScripts (scriptFiles) {
           const res = parseAst(PARSE_AST_TYPE.NORMAL, ast, [], item, outputItem)
           const fileDep = dependencyTree[item] || {}
           let resCode = res.code
-          resCode = await compileScriptFile(item, res.code)
+          resCode = await compileScriptFile(res.code)
           fs.ensureDirSync(path.dirname(outputItem))
-          resCode = Util.replaceContentEnv(resCode, projectConfig.env || {})
-          resCode = Util.replaceContentConstants(resCode, projectConfig.defineConstants || {})
           if (isProduction) {
             const uglifyPluginConfig = pluginsConfig.uglify || { enable: true }
             if (uglifyPluginConfig.enable) {
@@ -1756,7 +1801,7 @@ function watchFiles () {
         if (includeStyleJSPath.length) {
           includeStyleJSPath.forEach(async item => {
             let outputWXSSPath = null
-            outputWXSSPath = item.filePath.replace(path.extname(item.filePath), '.wxss')
+            outputWXSSPath = item.filePath.replace(path.extname(item.filePath), outputFilesTypes.STYLE)
             let modifySource = outputWXSSPath.replace(appPath + path.sep, '')
             modifySource = modifySource.split(path.sep).join('/')
             Util.printLog(Util.pocessTypeEnum.MODIFY, '样式文件', modifySource)
@@ -1780,7 +1825,7 @@ function watchFiles () {
             Util.printLog(Util.pocessTypeEnum.GENERATE, '样式文件', modifyOutput)
           })
         } else {
-          let outputWXSSPath = filePath.replace(path.extname(filePath), '.wxss')
+          let outputWXSSPath = filePath.replace(path.extname(filePath), outputFilesTypes.STYLE)
           let modifySource = outputWXSSPath.replace(appPath + path.sep, '')
           modifySource = modifySource.split(path.sep).join('/')
           Util.printLog(Util.pocessTypeEnum.MODIFY, '样式文件', modifySource)
@@ -1812,9 +1857,11 @@ function watchFiles () {
     })
 }
 
-async function build ({ watch }) {
-  process.env.TARO_ENV = Util.BUILD_TYPES.WEAPP
+async function build ({ watch, adapter }) {
+  process.env.TARO_ENV = adapter
   isProduction = !watch
+  buildAdapter = adapter
+  outputFilesTypes = Util.MINI_APP_FILES[buildAdapter]
   buildProjectConfig()
   copyFiles()
   appConfig = await buildEntry()