+++ packages/taro-qapp/src/api/interactive/index.js
@@ -1,21 +1,26 @@
 import prompt from '@system.prompt'
-import {generateUnSupportApi} from "../utils"
+import { generateUnSupportApi } from '../utils'
 
 export function showToast (options = {}) {
-
-  const { title = '', duration = 1500, success, complete } = options
+  const { title = '', duration = 1500, success, complete, fail } = options
   const res = { errMsg: 'showToast:ok' }
 
   return new Promise((resolve, reject) => {
-    prompt.showToast({
-      message: title,
-      duration: duration > 2000 ? 1 : 0
-    })
-    success && success(res)
-    complete && complete(res)
-    resolve(res)
+    try {
+      prompt.showToast({
+        message: title,
+        duration
+      })
+      success && success(res)
+      complete && complete(res)
+      resolve(res)
+    } catch (data) {
+      res.errMsg = 'showToast: error'
+      res.data = data
+      fail && fail(res)
+      reject(res)
+    }
   })
-
 }
 
 export function showModal (options = {}) {
@@ -116,16 +121,56 @@ export function showActionSheet (options = {}) {
   })
 }
 
-let unSupportApis = ['hideToast', 'showLoading', 'hideLoading']
+export function setNavigationBarTitle (options = {}) {
+  const { title = '', success, complete, fail } = options
+  const res = { errMsg: 'setNavigationBarTitle: ok' }
+
+  return new Promise((resolve, reject) => {
+    try {
+      this.$page.setTitleBar({text: title})
+      success && success(res)
+      complete && complete(res)
+      resolve(res)
+    } catch (data) {
+      res.errMsg = 'setNavigationBarTitle: error'
+      res.data = data
+      fail && fail(res)
+      reject(res)
+    }
+  })
+}
+
+export function setNavigationBarColor (options = {}) {
+  const { frontColor = '', backgroundColor = '', success, complete, fail } = options
+  const res = { errMsg: 'setNavigationBarColor: ok' }
+
+  return new Promise((resolve, reject) => {
+    try {
+      this.$page.setTitleBar({textColor: frontColor, backgroundColor})
+      success && success(res)
+      complete && complete(res)
+      resolve(res)
+    } catch (data) {
+      res.errMsg = 'setNavigationBarColor: error'
+      res.data = data
+      fail && fail(res)
+      reject(res)
+    }
+  })
+}
+
+let unSupportApis = ['hideToast', 'showLoading', 'hideLoading', 'showNavigationBarLoading', 'hideNavigationBarLoading']
 unSupportApis = generateUnSupportApi(
-  '快应用暂不支持storage的同步存取',
+  '快应用暂不支持Toast等隐藏方法',
   unSupportApis
 )
 
 const toast = {
   showToast,
   showModal,
-  showActionSheet
+  showActionSheet,
+  setNavigationBarTitle,
+  setNavigationBarColor
 }
 
 Object.assign(toast, unSupportApis)