+++ packages/taro-transformer-wx/src/index.ts
@@ -1,13 +1,14 @@
 import traverse, { Binding, NodePath } from 'babel-traverse'
 import generate from 'babel-generator'
 import * as fs from 'fs'
-import { Transformer } from './class'
 import { prettyPrint } from 'html'
+import { transform as parse } from 'babel-core'
+import * as ts from 'typescript'
+import { Transformer } from './class'
 import { setting, findFirstIdentifierFromMemberExpression, isContainJSXElement, codeFrameError } from './utils'
 import * as t from 'babel-types'
 import { DEFAULT_Component_SET, INTERNAL_SAFE_GET, TARO_PACKAGE_NAME, ASYNC_PACKAGE_NAME, REDUX_PACKAGE_NAME, IMAGE_COMPONENTS, INTERNAL_INLINE_STYLE, THIRD_PARTY_COMPONENTS, INTERNAL_GET_ORIGNAL } from './constant'
-import { transform as parse } from 'babel-core'
-import * as ts from 'typescript'
+import { Adapters, setAdapter } from './adapter'
 const template = require('babel-template')
 
 interface ENVS {
@@ -22,7 +23,8 @@ export interface Options {
   code: string,
   isTyped: boolean,
   isNormal?: boolean,
-  env?: ENVS
+  env?: ENVS,
+  adapter?: Adapters
 }
 
 function getIdsFromMemberProps (member: t.MemberExpression) {
@@ -151,6 +153,9 @@ interface TransformResult extends Result {
 }
 
 export default function transform (options: Options): TransformResult {
+  if (options.adapter) {
+    setAdapter(options.adapter)
+  }
   const code = options.isTyped
     ? ts.transpile(options.code, {
       jsx: ts.JsxEmit.Preserve,
@@ -185,7 +190,6 @@ export default function transform (options: Options): TransformResult {
     },
     plugins: [
       require('babel-plugin-transform-flow-strip-types'),
-      [require('babel-plugin-danger-remove-unused-import'), { ignore: ['@tarojs/taro', 'react', 'nervjs'] }],
       [require('babel-plugin-transform-define').default, {
         'process.env.TARO_ENV': taroEnv
       }]
@@ -202,6 +206,7 @@ export default function transform (options: Options): TransformResult {
   let mainClass!: NodePath<t.ClassDeclaration>
   let storeName!: string
   let renderMethod!: NodePath<t.ClassMethod>
+  let isImportTaro = false
   traverse(ast, {
     ClassDeclaration (path) {
       mainClass = path
@@ -390,6 +395,7 @@ export default function transform (options: Options): TransformResult {
       const source = path.node.source.value
       const names: string[] = []
       if (source === TARO_PACKAGE_NAME) {
+        isImportTaro = true
         path.node.specifiers.push(
           t.importSpecifier(t.identifier(INTERNAL_SAFE_GET), t.identifier(INTERNAL_SAFE_GET)),
           t.importSpecifier(t.identifier(INTERNAL_GET_ORIGNAL), t.identifier(INTERNAL_GET_ORIGNAL)),
@@ -424,6 +430,18 @@ export default function transform (options: Options): TransformResult {
       componentSourceMap.set(source, names)
     }
   })
+
+  if (!isImportTaro) {
+    ast.program.body.unshift(
+      t.importDeclaration([
+        t.importDefaultSpecifier(t.identifier('Taro')),
+        t.importSpecifier(t.identifier(INTERNAL_SAFE_GET), t.identifier(INTERNAL_SAFE_GET)),
+        t.importSpecifier(t.identifier(INTERNAL_GET_ORIGNAL), t.identifier(INTERNAL_GET_ORIGNAL)),
+        t.importSpecifier(t.identifier(INTERNAL_INLINE_STYLE), t.identifier(INTERNAL_INLINE_STYLE))
+      ], t.stringLiteral('@tarojs/taro'))
+    )
+  }
+
   if (!mainClass) {
     throw new Error('未找到 Taro.Component 的类定义')
   }